Updates Wordpress and plugins.

+++ web/app/plugins/wp-retina-2x/common/meow_admin.php
@@ -5,7 +5,7 @@ if ( !class_exists( 'Meow_Admin' ) ) {
 	class Meow_Admin {
 
 		public static $loaded = false;
-		public static $version = "0.1";
+		public static $version = "0.2";
 		public $prefix = null;
 		public $item = null;
 
@@ -17,11 +17,13 @@ if ( !class_exists( 'Meow_Admin' ) ) {
 					add_action( 'plugins_loaded', array( $this, 'plugins_loaded' ) );
 				}
 			}
-			if ( !empty( $prefix ) && !empty( $item ) && is_admin() ) {
+			if ( !empty( $prefix ) && !empty( $item ) ) {
 				$this->prefix = $prefix;
 				$this->item = $item;
-				add_action( 'update_option_' . $prefix . '_pro_serial', array( $this, 'serial_updated' ), 10, 2 );
-				add_action( 'admin_menu', array( $this, 'admin_menu_for_serialkey' ) );
+				if ( is_admin() ) {
+					add_action( 'update_option_' . $prefix . '_pro_serial', array( $this, 'serial_updated' ), 10, 2 );
+					add_action( 'admin_menu', array( $this, 'admin_menu_for_serialkey' ) );
+				}
 			}
 			Meow_Admin::$loaded = true;
 		}
@@ -253,15 +255,20 @@ if ( !class_exists( 'Meow_Admin' ) ) {
 				echo "</div>";
 			}
 			else if ( isset( $_GET['tool'] ) && $_GET['tool'] == 'error_log' ) {
+				$errorpath = ini_get( 'error_log' );
 				echo "<a href=\"javascript:history.go(-1)\">< Go back</a><br /><br />";
 				echo '<div id="error_log">';
-				echo "Now (auto-reload every 5 seconds):<br />[" . date( "d-M-Y H:i:s", time() ) . " UTC]<br /<br /><br />Errors (order by latest):";
-				$errorpath = ini_get( 'error_log' );
-				$errors = file_get_contents( $errorpath );
-				$errors = explode( "\n", $errors );
-				$errors = array_reverse( $errors );
-				$errors = implode( "<br />", $errors );
-				echo $errors;
+				if ( file_exists( $errorpath ) ) {
+					echo "Now (auto-reload every 5 seconds):<br />[" . date( "d-M-Y H:i:s", time() ) . " UTC]<br /<br /><br />Errors (order by latest):";
+					$errors = file_get_contents( $errorpath );
+					$errors = explode( "\n", $errors );
+					$errors = array_reverse( $errors );
+					$errors = implode( "<br />", $errors );
+					echo $errors;
+				}
+				else {
+					echo "The PHP Error Logs cannot be found. Please ask your hosting service for it.";
+				}
 				echo "</div>";
 
 			}
